Move highlight, new and init tests
